pyproject.toml
[project]
name = "hackathon-ibm-dia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.12.0",
    "fastapi[standard]>=0.121.0",
    "ibm-watsonx-ai>=1.4.4",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.3",
    "langchain-google-genai>=3.0.1",
]
